+++ src/config-examples/default-map-settings-example.js
@@ -29,6 +29,7 @@ const defaultMapSettings = {
   compressDataUrlSegment: true,
   autoFitHighlightedBounds: true,
   convertStopAfterRouteEndingToDestination: false,
+  useStopOptimization: false,
   acessibleModeActive: false,
   shownOnceTooltips: {},
   mapCenter: {lat: 49.510944, lng: 8.76709}, // Default center is at Heidelberg, Germany