+++ src/fragments/forms/settings/i18n/settings.i18n.it-it.js
@@ -51,6 +51,7 @@ export default {
     'resetShownOnceTooltips': 'Mostra di nuovo indicazioni',
     'convertStopAfterRouteEndingToDestination': 'Converti in destinazione la fermata dopo la fine del percorso',
     "forNauticalRoutes": 'Per le rotte nautiche, ad esempio',
-    "skipAllSegments": 'percorso rettilineo per tutti i segmenti'
+    "skipAllSegments": 'percorso rettilineo per tutti i segmenti',
+    "useStopOptimization": 'Ottimizza il percorso quando viene creata una sosta tramite trascinamento / salto'
   }
 }